Disable version/quota information for faulty accounts
When an account is configured incorrectly and files can't connect to the
backend server it does not make sense to allow clicking on the version
or quota information. Disable the showing of the popup and change the
mouse pointer to a normal one when hoevering over quota/version.

References: KFP-324