Use libsodium to encrypt backend configuration
The old encryption method of backend configurations used a hardcoded
algorithm and hardcoded password/iv which is not secure by default.
Change this to libsodium and by default set no password so the admin is
forced to generate one. The nonce is also saved in the backend
configuration and generated automatically when required.

References: KFP-428